chore: Change go module name and remove gcloud plugin (#3)
<!-- Describe what has changed in this PR -->
**What changed?**
Changed the module names to `github.com/cadence-workflow/shard-manager`
and `github.com/cadence-workflow/shard-manager/cmd/server`

Removed the gcloud module since we don't need it, so no point in
updating it

Disabled the breaking change GH action since it relies on the repo being
public to get the body.

<!-- Tell your future self why have you made these changes -->
**Why?**
We need different names since this is not cadence 

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

---------

Signed-off-by: Jakob Haahr Taankvist <jht@uber.com>